+++ tutorial/Delaunay/VanishSuperNodes.cpp
@@ -5,7 +5,7 @@ void VanishSuperNodes(vector<Xyc>&Z,vector<Nde>&N)
   long e, n=Z.size()-4;
 
 
-  for ( e = 1; e<N.size(); e++) 
+  for ( e = 1; e<(long)N.size(); e++) 
     if (N[e].a == 0 || N[e].b == 0 || N[e].c == 0 ||
 	N[e].a >  n || N[e].b >  n || N[e].c >  n ) {
       N[e].a = 0; N[e].b = 0; N[e].c = 0; N[e].A = 0; N[e].B = 0; N[e].C = 0;
@@ -18,7 +18,7 @@ void VanishSuperNodes(vector<Xyc>&Z,vector<Nde>&N)
   
   long i, j;
   j = 1;
-  for ( i = 1; i+j< N.size();) {
+  for ( i = 1; i+j< (long)N.size();) {
     if (N[i].a==0&& N[i].b ==0 && N[i].c == 0 &&
 	N[i].A==0&& N[i].B ==0 && N[i].C == 0 ){
       static Nde zero;
@@ -30,7 +30,7 @@ void VanishSuperNodes(vector<Xyc>&Z,vector<Nde>&N)
   for ( j=N.size(); i+1<j; i++)
     N.pop_back();
 
-  for ( i = 0; i<N.size(); i++)
+  for ( i = 0; i<(long)N.size(); i++)
     N[i].A = N[i].B = N[i].C = 0;
 }
